cargo-zaplib: Extract “build” to its own file
For consistency
