+++ zaplib/cargo-zaplib/src/main.rs
@@ -1,4 +1,6 @@
 #[cfg(not(target_arch = "wasm32"))]
+mod build;
+#[cfg(not(target_arch = "wasm32"))]
 mod cmd;
 #[cfg(not(target_arch = "wasm32"))]
 mod install_deps;
